[0.3.11] (破坏性更新) 令SQLUpdateAction返回的值为 Long 以适配自增主键大小范围。

+++ api/src/main/java/cc/carm/lib/easysql/api/action/PreparedSQLUpdateBatchAction.java
@@ -4,7 +4,7 @@ import cc.carm.lib.easysql.api.SQLAction;
 
 import java.util.List;
 
-public interface PreparedSQLUpdateBatchAction extends SQLAction<List<Integer>> {
+public interface PreparedSQLUpdateBatchAction extends SQLAction<List<Long>> {
 
     /**
      * 设定多组SQL语句中所有 ? 对应的参数
